+++ test_images/create.sh
@@ -1,44 +1,58 @@
 #!/bin/sh
 
 echo "this is version 1.0.0 for image a" > file1
-docker build -f DockerfileA_1 -t testimagea:1.0.0 -t testimagea:1.0 -t testimagea:1 -t testimagea:mainline -t testimagea:latest .
-#docker push testimagea:1.0.0
-#docker push testimagea:1.0
-#docker push testimagea:1
-#docker push testimagea:latest
-#docker push testimagea:mainline
+docker build -f DockerfileA_1 -t menedev/testimagea:1.0.0 -t menedev/testimagea:1.0 -t menedev/testimagea:1 -t menedev/testimagea:mainline -t menedev/testimagea:latest .
+docker push menedev/testimagea:1.0.0
+docker push menedev/testimagea:1.0
+docker push menedev/testimagea:1
+docker push menedev/testimagea:latest
+docker push menedev/testimagea:mainline
 
 echo "this is version 1.0.1 for image a" > file1
-docker build -f DockerfileA_1 -t testimagea:1.0.1 -t testimagea:1.0 -t testimagea:1 -t testimagea:mainline -t testimagea:latest .
-#docker push testimagea:1.0.1
-#docker push testimagea:1.0
-#docker push testimagea:1
-#docker push testimagea:latest
-#docker push testimagea:mainline
+docker build -f DockerfileA_1 -t menedev/testimagea:1.0.1 -t menedev/testimagea:1.0 -t menedev/testimagea:1 -t menedev/testimagea:mainline -t menedev/testimagea:latest .
+docker push menedev/testimagea:1.0.1
+docker push menedev/testimagea:1.0
+docker push menedev/testimagea:1
+docker push menedev/testimagea:latest
+docker push menedev/testimagea:mainline
 
 echo "this is version 1.1.0 for image a" > file1
-docker build -f DockerfileA_1 -t testimagea:1.1.0 -t testimagea:1.1 -t testimagea:1 -t testimagea:mainline -t testimagea:latest .
-#docker push testimagea:1.1.0
-#docker push testimagea:1.1
-#docker push testimagea:1
-#docker push testimagea:latest
-#docker push testimagea:mainline
+docker build -f DockerfileA_1 -t menedev/testimagea:1.1.0 -t menedev/testimagea:1.1 -t menedev/testimagea:1 -t menedev/testimagea:mainline -t menedev/testimagea:latest .
+docker push menedev/testimagea:1.1.0
+docker push menedev/testimagea:1.1
+docker push menedev/testimagea:1
+docker push menedev/testimagea:latest
+docker push menedev/testimagea:mainline
 
 echo "this is version 1.1.1 for image a" > file1
-docker build -f DockerfileA_1 -t testimagea:1.1.1 -t testimagea:1.1 -t testimagea:1 -t testimagea:mainline -t testimagea:latest .
-#docker push testimagea:1.1.1
-#docker push testimagea:1.1
-#docker push testimagea:1
-#docker push testimagea:latest
-#docker push testimagea:mainline
+docker build -f DockerfileA_1 -t menedev/testimagea:1.1.1 -t menedev/testimagea:1.1 -t menedev/testimagea:1 -t menedev/testimagea:mainline -t menedev/testimagea:latest .
+docker push menedev/testimagea:1.1.1
+docker push menedev/testimagea:1.1
+docker push menedev/testimagea:1
+docker push menedev/testimagea:latest
+docker push menedev/testimagea:mainline
 
 echo "this is version 2.0.0 for image a" > file1
-docker build -f DockerfileA_1 -t testimagea:2.0.0 -t testimagea:2.0 -t testimagea:2 -t testimagea:edge -t testimagea:registry-only -t testimagea:latest .
-#docker push testimagea:2.0.0
-#docker push testimagea:2.0
-#docker push testimagea:2
-#docker push testimagea:latest
-#docker push testimagea:edge
-#docker push testimagea:registry-only
-#
-#docker rmi testimagea:registry-only
+docker build -f DockerfileA_1 -t menedev/testimagea:2.0.0 -t menedev/testimagea:2.0 -t menedev/testimagea:2 -t menedev/testimagea:edge -t menedev/testimagea:registry-only -t menedev/testimagea:latest .
+docker push menedev/testimagea:2.0.0
+docker push menedev/testimagea:2.0
+docker push menedev/testimagea:2
+docker push menedev/testimagea:latest
+docker push menedev/testimagea:edge
+docker push menedev/testimagea:registry-only
+
+docker pull menedev/testimagea:1
+docker pull menedev/testimagea:1.0
+docker pull menedev/testimagea:1.0.0
+docker pull menedev/testimagea:1.0.1
+docker pull menedev/testimagea:1.1
+docker pull menedev/testimagea:1.1.1
+docker pull menedev/testimagea:2
+docker pull menedev/testimagea:2.0
+docker pull menedev/testimagea:2.0.0
+docker pull menedev/testimagea:edge
+docker pull menedev/testimagea:latest
+docker pull menedev/testimagea:mainline
+docker pull menedev/testimagea:registry-only
+
+docker rmi menedev/testimagea:registry-only